Add hs-bindgen-libclang with build-type: Configure
Resolves #78.

Having C-dependencies in a separate package work-around unclear
behavior of `build-type: Configure` which seems to add extra definitions
to the main library only (`a-lib.buildinfo`; possibly we can have extra
buildinfos for other components too, but the file naming scheme is
mysterious).

On the other hand, separate package for `libclang` bindings might be a
good idea anyway. It's a separate self-contained functionality,
and e.g. `clang-tutorial` depends on that only (and could be pulled into
separate package as well, not to pollute the main thing).