Add ClangArgsConfig
There are two reasons for using a separate configuration type:

* Clang arguments are determined based on configuration impurely.
  Arguments may be added depending on `BINDGEN_CLANG_EXTRA_ARGS`
  environment variables.  Builtin include directory arguments may
  also be added.  Both should be done once (during "boot").

* `Clang.Args` implemented many `hs-bindgen` features/limitations.  We
  plan on making the `clang` package separate from `hs-bindgen`, so
  those features/limitations should be implemented in `hs-bindgen`.

The `ClangArgs` type is changed to be a trivial `newtype` wrapper around
raw arguments (`[String]`).

`CStandard` is still defined in `Clang.Args`, as it is *not* specific to
`hs-bindgen`.  Function `getStdClangArg` gets the appropriate Clang
option, depending on the version of Clang.  Sum type
`Gnu = DisableGnu | EnableGnu` is added to support this.  Note that this
functionality is used in the `c-expr-runtime` tests, which do not depend
on `hs-bindgen`.

Following our configuration convention, `ClangArgsConfig` is defined in
`HsBindgen.Config.ClangArgs`.  The `BuiltinIncDirConfig` type is moved
into this module, as this configuration is part of `ClangArgsConfig`.

The `BINDGEN_EXTRA_CLANG_ARGS` functionality is moved to
`HsBindgen.Clang.ExtraClangArgs`, alongside the `BuiltingConfigDir`
module.  Note that I ended up refactoring `getExtraClangArgs` because
there was a bug: the environment variable for the target was always
reported in the trace message even when falling back to
`BINDGEN_EXTRA_CLANG_ARGS`.

Resolving the Clang arguments with processing of
`BINDGEN_CLANG_EXTRA_ARGS` environment variables and builtin include
directory configuration is a common action, needed in development
commands that cannot rely on the whole `boot` configuration/artifact.
I implemented it in `HsBindgen.Boot` as function `getClangArgs`.  This
function is used within `boot` as well as in the development commands
that need it.  Note that exception `InvalidClangArgs` was previously
handled deep inside of our `clang` library, using `throwIO`.  We can
now handle it in this `getClangArgs` function.  I kept `throwIO` for
now, but we can change this if desired.