Drop `runFold`
The contract of `runFold` promised

```hs
--- If the fold has an associated exception handler, it will get a chance to
--- catch any exceptions thrown. See 'foldWithHandler' for detailed discussion.
```

but that is a contract it could not keep: after all, the whole point for introducing `foldWithHandler` in the first place was that we need to treat these handlers very carefully, simply using `handle` does not work.

It turns out we don't actually _need_ `runFold`, so deleted it altogether; I
think this clarifies the situation on the `hs-bindgen` side also (see
discussion of `topLevelFold` there).